requirements.txt
pandas==2.3.3
openpyxl==3.1.5
pyyaml>=6.0
xlrd==2.0.2
xlwings==0.33.16
numpy==2.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
et_xmlfile==2.0.0
lxml==6.0.2
psutil==7.1.1

# Streamlit app (visual data tool)
streamlit>=1.28.0
plotly>=5.18.0
pyarrow>=14.0.0